style: eslint fixes
